+++ apps/docs/src/components/colors/colors.css.ts
@@ -1,5 +1,5 @@
 import { globalStyle, style } from "@vanilla-extract/css";
-import tokens from "guds-tokens/output/tokens.json";
+import tokens from "gyldendal-tokens/output/tokens.json";
 import typography from "gu-ds-css/output/typography.json";
 
 export const grid = style({