+++ packages/gu-ds-components/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Added progress bar components
 - Updated dependencies
-  - guds-tokens@0.0.3
+  - gyldendal-tokens@0.0.3
 
 ## 0.0.2